feat(game) : 3자리 난수 생성
