+++ src/main/java/baseball/Player.java
@@ -1,0 +1,5 @@
+package baseball;
+
+public class Player {
+    
+}
