+++ src/main/java/utils/RandomUtils.java
@@ -24,18 +24,4 @@ public class RandomUtils {
         return startInclusive + RANDOM.nextInt(endInclusive - startInclusive + 1);
     }
 
-    public static String makeGameNumber(){
-        String gameNumber = "";
-        int countLength = 0;
-
-        while (countLength < 3) {
-            int randomNumber = RandomUtils.nextInt(1, 9);
-
-            if (!gameNumber.contains(String.valueOf(randomNumber))) {
-                gameNumber += randomNumber;
-                countLength++;
-            }
-        }
-        return gameNumber;
-    }
 }
